policy nits: name the contract bound, and one comparison instead of two
The manifest's output-contract length bound was a bare `64` beside a
named document bound; it gets a name.

`evaluate_where` built an `Option` with `then_some` only to unwrap it in
the next line — the condition can just be the `if`.

The content binding hashed lengths as bare `usize`, the same
architecture-dependent encoding the policy hash just lost. This hash IS
compared across machines — a relay recomputes it locally and matches it
against one a vault produced — so it takes the same eight-byte rule.

`from_edge_auth`'s `dead_code` allow re-checked post-BYO: still only
tests and the compile-fail pins call it, so the allow stays, now with a
line saying what would retire it.

+++ crates/oneiron/src/policy_model/binding.rs
@@ -95,7 +95,11 @@ fn hash_binding_opt_str(hasher: &mut Sha256, label: &str, value: Option<&str>) {
 fn hash_binding_str(hasher: &mut Sha256, label: &str, value: &str) {
     hasher.update(label.as_bytes());
     hasher.update([0]);
-    hasher.update(value.len().to_be_bytes());
+    // Eight bytes on every architecture. A relay recomputes this hash locally
+    // and compares it against one a VAULT produced, so a bare `usize` — four
+    // bytes on a 32-bit target, eight on a 64-bit one — would make the two
+    // machines disagree about identical content. Same rule as the policy hash.
+    hasher.update((value.len() as u64).to_be_bytes());
     hasher.update(value.as_bytes());
     hasher.update([0xff]);
 }
